Moved device tests into handle module
