+++ src/device/lib.rs
@@ -361,48 +361,3 @@ impl<D: Device> MapFactory for D {
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use std::mem;
-    use std::marker::PhantomData;
-    use super::{handle, BufferInfo, BufferUsage};
-
-    #[derive(Clone, Debug, PartialEq)]
-    enum TestResources {}
-    impl super::Resources for TestResources {
-        type Buffer = ();
-        type ArrayBuffer = ();
-        type Shader = ();
-        type Program = ();
-        type FrameBuffer = ();
-        type Surface = ();
-        type Texture = ();
-        type Sampler = ();
-    }
-
-    fn mock_buffer<T>(len: usize) -> handle::Buffer<TestResources, T> {
-        handle::Buffer {
-            raw: unsafe { handle::RawBuffer::new(
-                (),
-                BufferInfo {
-                    usage: BufferUsage::Static,
-                    size: mem::size_of::<T>() * len,
-                },
-            ) },
-            phantom_t: PhantomData,
-        }
-    }
-
-    #[test]
-    fn test_buffer_len() {
-        assert_eq!(mock_buffer::<u8>(8).len(), 8);
-        assert_eq!(mock_buffer::<u16>(8).len(), 8);
-    }
-
-    #[test]
-    #[should_fail]
-    fn test_buffer_zero_len() {
-        let _ = mock_buffer::<()>(0).len();
-    }
-}
